requirements.txt
numpy
nuscenes-devkit
Pillow
torch
scipy